orderId to Long
The type of the orderId is updated to Long

+++ src/test/java/com/payu/sdk/reporting/ReportingApiIntegrationTest.java
@@ -71,7 +71,7 @@ public class ReportingApiIntegrationTest {
 	/**
 	 * The created order id
 	 */
-	private static Integer orderId;
+	private static Long orderId;
 
 	/**
 	 * the created transactionId
